+++ src/android/io/sentry/SentryCordova.java
@@ -62,21 +62,25 @@ public class SentryCordova extends CordovaPlugin {
 
         break;
       case "captureEnvelope":
-        String headerString = null;
-        String payloadString = null;
-        String payloadType = "event";
-
-        if (!args.isNull(0)) {
-          headerString = args.getString(0);
+        if (args.isNull(0)) {
+          break;
         }
-        if (!args.isNull(1)) {
-          payloadString = args.getString(1);
-        }
-        if (!args.isNull(2)) {
-          payloadType = args.getString(2);
+        JSONArray envelopeRaw = args.getJSONObject(0).getJSONArray("envelope");
+        byte[] bytes = new byte[envelopeRaw.length()];
+        for (int i = 0; i < bytes.length; i++) {
+          bytes[i] = (byte) envelopeRaw.getInt(i);
         }
 
-        captureEnvelope(headerString, payloadString, payloadType, callbackContext);
+        final String outboxPath = HubAdapter.getInstance().getOptions().getOutboxPath();
+
+        final File installation = new File(outboxPath, UUID.randomUUID().toString());
+
+        try (FileOutputStream out = new FileOutputStream(installation)) {
+          out.write(bytes);
+          logger.info("Successfully captured envelope.");
+        } catch (Exception e) {
+          logger.info("Error writing envelope.");
+        }
 
         break;
       case "setUser":