+++ src/js/integrations/index.ts
@@ -1,1 +1,3 @@
 export { Cordova } from './cordova';
+export { EventOrigin } from './eventorigin';
+export { SdkInfo } from './sdkinfo';