+++ src/js/sdk.ts
@@ -1,18 +1,15 @@
-import { defaultIntegrations as defaultBrowserIntegrations } from '@sentry/browser';
-import { Hub, initAndBind, makeMain } from '@sentry/core';
+import type { BrowserOptions } from '@sentry/browser';
+import { defaultIntegrations, init as browserInit } from '@sentry/browser';
+import { Hub, makeMain } from '@sentry/core';
 import { getGlobalObject } from '@sentry/utils';
 
-import { CordovaOptions } from './backend';
-import { CordovaClient } from './client';
-import { Cordova } from './integrations';
+import { Cordova, EventOrigin, SdkInfo } from './integrations';
+import type { CordovaOptions } from './options';
 import { CordovaScope } from './scope';
+import { makeCordovaTransport } from './transports/cordova';
 import { NATIVE } from './wrapper';
-
-const DEFAULT_INTEGRATIONS = [...defaultBrowserIntegrations, new Cordova()];
-
 const DEFAULT_OPTIONS: CordovaOptions = {
   enableNative: true,
-  defaultIntegrations: DEFAULT_INTEGRATIONS,
   enableAutoSessionTracking: true,
   enableNdkScopeSync: false,
   attachThreads: false,
@@ -21,21 +18,63 @@ const DEFAULT_OPTIONS: CordovaOptions = {
 /**
  * Inits the SDK
  */
-export function init(_options: Partial<CordovaOptions>): void {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function init(options: Partial<CordovaOptions>): void {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, deprecation/deprecation
   const window = getGlobalObject<{ SENTRY_RELEASE?: { id?: string } }>();
 
-  const options = {
+  const finalOptions = {
+    enableAutoSessionTracking: true,
+    enableOutOfMemoryTracking: true,
     ...DEFAULT_OPTIONS,
     release: window?.SENTRY_RELEASE?.id,
-    ..._options,
+    ...options,
   };
+
+  if (finalOptions.enabled === false ||
+    NATIVE.platform === 'browser') {
+    finalOptions.enableNative = false;
+    finalOptions.enableNativeNagger = false;
+  } else {
+    // keep the original value if user defined it.
+    if (finalOptions.enableNativeNagger === undefined) {
+      finalOptions.enableNativeNagger = true;
+    }
+    if (finalOptions.enableNative === undefined) {
+      finalOptions.enableNative = true;
+    }
+  }
 
   // Initialize a new hub using our scope with native sync
   const cordovaHub = new Hub(undefined, new CordovaScope());
   makeMain(cordovaHub);
 
-  initAndBind(CordovaClient, options);
+  finalOptions.defaultIntegrations = [
+    ...defaultIntegrations,
+    new SdkInfo(),
+    new EventOrigin(),
+    new Cordova(),
+  ];
+
+  if (!options.transport && finalOptions.enableNative) {
+    finalOptions.transport = options.transport || makeCordovaTransport;
+  }
+
+  const browserOptions = {
+    ...finalOptions,
+    autoSessionTracking:
+      NATIVE.platform === 'browser' && finalOptions.enableAutoSessionTracking,
+  } as BrowserOptions;
+
+  const mobileOptions = {
+    ...finalOptions,
+    enableAutoSessionTracking:
+      NATIVE.platform !== 'browser' && finalOptions.enableAutoSessionTracking,
+  } as CordovaOptions;
+
+  // We first initialize the NATIVE SDK to avoid the Javascript SDK to invoke any
+  // feature from the NATIVE SDK without the options being set.
+  void NATIVE.startWithOptions(mobileOptions);
+  browserInit(browserOptions);
 }
 
 /**