+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+### Dependencies
+
+- Bump Sentry JavaScript SDK to `7.34.0` ([#302](https://github.com/getsentry/sentry-cordova/pull/302))
+  - [changelog](https://github.com/getsentry/sentry-javascript/releases/tag/7.34.0)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/6.1.0...7.34.0)
+- build(android): Bump Android SDK to 6.11.0 ([#302](https://github.com/getsentry/sentry-cordova/pull/302))
+  - [changelog](https://github.com/getsentry/sentry-java/releases/tag/6.11.0)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.4.1...6.11.0)
+- build(ios): Bump `sentry-cocoa` to 7.27.1 ([#302](https://github.com/getsentry/sentry-cordova/pull/302))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/releases/tag/7.27.1)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/6.2.1...7.27.1)
+
 ## 1.0.4
 
 ### Fixes